dbsp: Factor out common code for cursors in `dbsp::trace::ord::fallback`.
This code was copy-pasted privately into two modules.  It could be useful
more broadly (and it will be used more broadly in upcoming commits), so
this moves it into `dbsp::trace::cursor` and makes it public.
