+++ crates/dbsp/src/trace/ord/fallback/indexed_wset.rs
@@ -6,6 +6,7 @@ use crate::{
     storage::file::reader::Error as ReaderError,
     time::AntichainRef,
     trace::{
+        cursor::DelegatingCursor,
         layers::OrdOffset,
         ord::{
             file::indexed_wset_batch::{
@@ -338,7 +339,7 @@ where
     type Val = V;
     type Time = ();
     type R = R;
-    type Cursor<'s> = FallbackIndexedWSetCursor<'s, K, V, R>
+    type Cursor<'s> = DelegatingCursor<'s, K, V, (), R>
     where
         V: 's;
 
@@ -348,7 +349,10 @@ where
 
     #[inline]
     fn cursor(&self) -> Self::Cursor<'_> {
-        FallbackIndexedWSetCursor::new(self)
+        DelegatingCursor(match &self.inner {
+            Inner::Vec(vec) => Box::new(vec.cursor()),
+            Inner::File(file) => Box::new(file.cursor()),
+        })
     }
 
     #[inline]
@@ -562,166 +566,6 @@ where
             MergerInner::ToFile(merger) => merger.size_of_children(context),
             MergerInner::ToVec(merger) => merger.size_of_children(context),
         }
-    }
-}
-
-trait ClonableCursor<'s, K, V, R>: Cursor<K, V, (), R> + Debug
-where
-    K: ?Sized,
-    V: ?Sized,
-    R: ?Sized,
-{
-    fn clone_boxed(&self) -> Box<dyn ClonableCursor<'s, K, V, R> + 's>;
-}
-
-impl<'s, K, V, R, C> ClonableCursor<'s, K, V, R> for C
-where
-    K: ?Sized,
-    V: ?Sized,
-    R: ?Sized,
-    C: Cursor<K, V, (), R> + Debug + Clone + 's,
-{
-    fn clone_boxed(&self) -> Box<dyn ClonableCursor<'s, K, V, R> + 's> {
-        Box::new(self.clone())
-    }
-}
-/// A cursor for navigating a single layer.
-#[derive(Debug, SizeOf)]
-pub struct FallbackIndexedWSetCursor<'s, K, V, R>(Box<dyn ClonableCursor<'s, K, V, R> + 's>)
-where
-    K: DataTrait + ?Sized,
-    V: DataTrait + ?Sized,
-    R: WeightTrait + ?Sized;
-
-impl<'s, K, V, R> Clone for FallbackIndexedWSetCursor<'s, K, V, R>
-where
-    K: DataTrait + ?Sized,
-    V: DataTrait + ?Sized,
-    R: WeightTrait + ?Sized,
-{
-    fn clone(&self) -> Self {
-        Self(self.0.clone_boxed())
-    }
-}
-
-impl<'s, K, V, R> FallbackIndexedWSetCursor<'s, K, V, R>
-where
-    K: DataTrait + ?Sized,
-    V: DataTrait + ?Sized,
-    R: WeightTrait + ?Sized,
-{
-    fn new(wset: &'s FallbackIndexedWSet<K, V, R>) -> Self {
-        Self(match &wset.inner {
-            Inner::Vec(vec) => Box::new(vec.cursor()),
-            Inner::File(file) => Box::new(file.cursor()),
-        })
-    }
-}
-
-impl<'s, K, V, R> Cursor<K, V, (), R> for FallbackIndexedWSetCursor<'s, K, V, R>
-where
-    K: DataTrait + ?Sized,
-    V: DataTrait + ?Sized,
-    R: WeightTrait + ?Sized,
-{
-    fn weight_factory(&self) -> &'static dyn Factory<R> {
-        self.0.weight_factory()
-    }
-
-    fn key(&self) -> &K {
-        self.0.key()
-    }
-
-    fn val(&self) -> &V {
-        self.0.val()
-    }
-
-    fn map_times(&mut self, logic: &mut dyn FnMut(&(), &R)) {
-        self.0.map_times(logic)
-    }
-
-    fn map_times_through(&mut self, upper: &(), logic: &mut dyn FnMut(&(), &R)) {
-        self.0.map_times_through(upper, logic)
-    }
-
-    fn map_values(&mut self, logic: &mut dyn FnMut(&V, &R)) {
-        self.0.map_values(logic)
-    }
-
-    fn weight(&mut self) -> &R {
-        self.0.weight()
-    }
-
-    fn key_valid(&self) -> bool {
-        self.0.key_valid()
-    }
-
-    fn val_valid(&self) -> bool {
-        self.0.val_valid()
-    }
-
-    fn step_key(&mut self) {
-        self.0.step_key()
-    }
-
-    fn step_key_reverse(&mut self) {
-        self.0.step_key_reverse()
-    }
-
-    fn seek_key(&mut self, key: &K) {
-        self.0.seek_key(key)
-    }
-
-    fn seek_key_with(&mut self, predicate: &dyn Fn(&K) -> bool) {
-        self.0.seek_key_with(predicate)
-    }
-
-    fn seek_key_with_reverse(&mut self, predicate: &dyn Fn(&K) -> bool) {
-        self.0.seek_key_with_reverse(predicate)
-    }
-
-    fn seek_key_reverse(&mut self, key: &K) {
-        self.0.seek_key_reverse(key)
-    }
-
-    fn step_val(&mut self) {
-        self.0.step_val()
-    }
-
-    fn seek_val(&mut self, val: &V) {
-        self.0.seek_val(val)
-    }
-
-    fn seek_val_with(&mut self, predicate: &dyn Fn(&V) -> bool) {
-        self.0.seek_val_with(predicate)
-    }
-
-    fn rewind_keys(&mut self) {
-        self.0.rewind_keys()
-    }
-
-    fn fast_forward_keys(&mut self) {
-        self.0.fast_forward_keys()
-    }
-
-    fn rewind_vals(&mut self) {
-        self.0.rewind_vals()
-    }
-
-    fn step_val_reverse(&mut self) {
-        self.0.step_val_reverse()
-    }
-
-    fn seek_val_reverse(&mut self, val: &V) {
-        self.0.seek_val_reverse(val)
-    }
-
-    fn seek_val_with_reverse(&mut self, predicate: &dyn Fn(&V) -> bool) {
-        self.0.seek_val_with_reverse(predicate)
-    }
-
-    fn fast_forward_vals(&mut self) {
-        self.0.fast_forward_vals()
     }
 }
